requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
chromadb==0.4.15
openai==0.28.1
python-dotenv==1.0.0
plotly==5.17.0
altair==5.1.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlite3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
xlrd==2.0.1
seaborn==0.12.2
matplotlib==3.7.2
scipy==1.11.3
nltk==3.8.1
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
datasets==2.14.5
